# docs/swagger.yaml
basePath: /
...
host: localhost:9999
info:
  contact: {}
  description: Authentication Service
  title: Autentico OIDC
  version: "1.0"
paths:
...
  /oauth2/signup:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Renders the signup page (GET) or processes a new user registration
        (POST).
      parameters:
      - description: Desired username
        in: formData
        name: username
        type: string
      - description: Password
        in: formData
        name: password
        type: string
      - description: Confirm password
        in: formData
        name: confirm_password
        type: string
      - description: Email address
        in: formData
        name: email
        type: string
      - description: Redirect URI
        in: formData
        name: redirect_uri
        type: string
      - description: OAuth2 state
        in: formData
        name: state
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Signup form (GET)
          schema:
            type: string
        "302":
          description: Redirect back to client with code (POST)
          schema:
            type: string
      summary: User signup
      tags:
      - signup
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Renders the signup page (GET) or processes a new user registration
        (POST).
      parameters:
      - description: Desired username
        in: formData
        name: username
        type: string
      - description: Password
        in: formData
        name: password
        type: string
      - description: Confirm password
        in: formData
        name: confirm_password
        type: string
      - description: Email address
        in: formData
        name: email
        type: string
      - description: Redirect URI
        in: formData
        name: redirect_uri
        type: string
      - description: OAuth2 state
        in: formData
        name: state
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Signup form (GET)
          schema:
            type: string
        "302":
          description: Redirect back to client with code (POST)
          schema:
            type: string
      summary: User signup
      tags:
      - signup
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"